Added Controller interface Possibly improved readability with sub-packages and code formatting
