+++ src/main/java/io/molnarsandor/trelloclone/collaborator/CollaboratorRepository.java
@@ -1,5 +1,6 @@
 package io.molnarsandor.trelloclone.collaborator;
 
+import io.molnarsandor.trelloclone.collaborator.model.CollaboratorEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
