+++ src/main/java/io/molnarsandor/trelloclone/project/model/ProjectEntity.java
@@ -1,12 +1,12 @@
-package io.molnarsandor.trelloclone.project;
+package io.molnarsandor.trelloclone.project.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import io.molnarsandor.trelloclone.collaborator.CollaboratorEntity;
-import io.molnarsandor.trelloclone.project_task.BacklogEntity;
-import io.molnarsandor.trelloclone.user.UserEntity;
+import io.molnarsandor.trelloclone.collaborator.model.CollaboratorEntity;
+import io.molnarsandor.trelloclone.project_task.model.BacklogEntity;
+import io.molnarsandor.trelloclone.user.model.UserEntity;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
