+++ src/main/java/io/molnarsandor/trelloclone/project_task/BacklogRepository.java
@@ -1,5 +1,6 @@
 package io.molnarsandor.trelloclone.project_task;
 
+import io.molnarsandor.trelloclone.project_task.model.BacklogEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
