+++ src/main/java/io/molnarsandor/trelloclone/project_task/model/ProjectTaskDTO.java
@@ -1,4 +1,4 @@
-package io.molnarsandor.trelloclone.project_task;
+package io.molnarsandor.trelloclone.project_task.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
