+++ src/main/java/io/molnarsandor/trelloclone/user/UserService.java
@@ -1,5 +1,7 @@
 package io.molnarsandor.trelloclone.user;
 
+import io.molnarsandor.trelloclone.user.model.UserEntity;
+
 public interface UserService {
 
     UserEntity loadUserById(Long id);
