+++ src/main/java/io/molnarsandor/trelloclone/user/model/UserEntity.java
@@ -1,9 +1,9 @@
-package io.molnarsandor.trelloclone.user;
+package io.molnarsandor.trelloclone.user.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import io.molnarsandor.trelloclone.project.ProjectEntity;
+import io.molnarsandor.trelloclone.project.model.ProjectEntity;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
